add slowdown estimation mechanism.
Stats keeps track of the error rate in each epoch. Error rate is compared against the IPC of the application when it is running alone.

+++ hring/src/Controller/Global_RR/Controller_Global_RR.cs
@@ -72,7 +72,7 @@ namespace ICSimulator
             {
                 //prev_MPKI=MPKI[i];
                 //MPKI[i]=0.0;
-                num_ins_last_epoch[i] = Simulator.stats.insns_persrc[i].Count;
+				num_ins_last_epoch[i] = (ulong)Simulator.stats.insns_persrc[i].Count;
                 L1misses[i]=0;
             }            
         }
