+++ hring/src/Controller/Global_RR/Controller_Three_Level_Cluster.cs
@@ -133,7 +133,7 @@ namespace ICSimulator
                 Array.Clear(last_free_nodes,0,last_free_nodes.Length);
                 foreach (int node in nodes)
                 {
-                    ins_free[node]=Simulator.stats.insns_persrc[node].Count;
+					ins_free[node]=(ulong)Simulator.stats.insns_persrc[node].Count;
                     last_free_nodes[node]=1;
 
                     setThrottleRate(node,false);
@@ -224,10 +224,10 @@ namespace ICSimulator
             {
                 /* Calculate IPC during the last free-inject and throtttled interval */
                 if(ins_throttled[node]==0)
-                    ins_throttled[node]=Simulator.stats.insns_persrc[node].Count;
+					ins_throttled[node]=(ulong)Simulator.stats.insns_persrc[node].Count;
                 if(last_free_nodes[node]==1)
                 {
-                    ulong ins_retired=Simulator.stats.insns_persrc[node].Count-ins_free[node];
+					ulong ins_retired=(ulong)Simulator.stats.insns_persrc[node].Count-ins_free[node];
                     ipc_free[node]+=(double)ins_retired/Config.interval_length;
 #if DEBUG_CLUSTER
                     Console.Write("Free calc: Node ");
@@ -239,10 +239,10 @@ namespace ICSimulator
                 }
                 else
                 {
-                    ulong ins_retired=Simulator.stats.insns_persrc[node].Count-ins_throttled[node];
+					ulong ins_retired=(ulong)Simulator.stats.insns_persrc[node].Count-ins_throttled[node];
                     ipc_throttled[node]+=(double)ins_retired/Config.interval_length;
                 }
-                ins_throttled[node]=Simulator.stats.insns_persrc[node].Count;
+				ins_throttled[node]=(ulong)Simulator.stats.insns_persrc[node].Count;
             }
         }
