+++ hring/src/Controller/Global_RR/Controller_Three_Level_Cluster_NoTrigger.cs
@@ -157,7 +157,7 @@ namespace ICSimulator
         {
             for (int i = 0; i < Config.N; i++)
             {
-                num_ins_last_epoch[i] = Simulator.stats.every_insns_persrc[i].Count;
+				num_ins_last_epoch[i] = (ulong)Simulator.stats.every_insns_persrc[i].Count;
                 L1misses[i]=0;
             }            
         }
